requirements.txt
ccxt==4.*
pandas==2.*
numpy==1.*
PyYAML==6.*
python-dotenv==1.*
requests==2.*
ta==0.11.*
matplotlib==3.*
pytest==8.*
